Use loaded shift signups in admin page builder
