+++ src/Humans.Web/Models/Shifts/ShiftAdminPageBuilder.cs
@@ -1,7 +1,9 @@
 using Humans.Application;
+using Humans.Application.Interfaces.Governance;
 using Humans.Application.Interfaces.Shifts;
 using Humans.Application.Interfaces.Teams;
 using Humans.Application.Interfaces.Users;
+using Humans.Domain.Constants;
 using Humans.Domain.Entities;
 using Humans.Domain.Enums;
 using NodaTime;
@@ -19,7 +21,7 @@ public sealed record ShiftAdminPageRequest(
 
 public sealed class ShiftAdminPageBuilder(
     IShiftManagementService shiftManagement,
-    IShiftSignupService signupService,
+    IMembershipCalculator membership,
     IUserServiceRead userService,
     ITeamServiceRead teamService)
 {
@@ -72,16 +74,28 @@ public sealed class ShiftAdminPageBuilder(
             foreach (var shift in rota.Shifts)
             {
                 totalSlots += shift.MaxVolunteers;
-                var shiftSignups = await signupService.GetByShiftAsync(shift.Id);
-                confirmedCount += shiftSignups.Count(s => s.Status == SignupStatus.Confirmed);
-                pendingSignups.AddRange(shiftSignups.Where(s => s.Status == SignupStatus.Pending));
+                confirmedCount += shift.ShiftSignups.Count(s => s.Status == SignupStatus.Confirmed);
+                pendingSignups.AddRange(shift.ShiftSignups.Where(s => s.Status == SignupStatus.Pending));
             }
         }
 
         if (incompleteOnboarding)
-            pendingSignups = (await signupService.FilterToIncompleteOnboardingAsync(pendingSignups)).ToList();
+            pendingSignups = (await FilterToIncompleteOnboardingAsync(pendingSignups)).ToList();
 
         return (pendingSignups, totalSlots, confirmedCount);
+    }
+
+    private async Task<IReadOnlyList<ShiftSignup>> FilterToIncompleteOnboardingAsync(
+        IReadOnlyList<ShiftSignup> signups)
+    {
+        if (signups.Count == 0) return signups;
+
+        var userIds = signups.Select(s => s.UserId).Distinct().ToList();
+        var withConsents = await membership.GetUsersWithAllRequiredConsentsForTeamAsync(
+            userIds,
+            SystemTeamIds.Volunteers);
+
+        return signups.Where(s => !withConsents.Contains(s.UserId)).ToList();
     }
 
     private async Task<Dictionary<Guid, VolunteerBadgesViewModel>> LoadProfilesAsync(
