+++ tests/Humans.Application.Tests/Architecture/Baselines/ApplicationServiceEntityReadReturns.baseline.txt
@@ -21,7 +21,6 @@ Humans.Application.Interfaces.Shifts.IShiftManagementService.GetShiftProfileAsyn
 Humans.Application.Interfaces.Shifts.IShiftManagementService.GetShiftsByRotaAsync:Humans.Domain.Entities.Shift
 Humans.Application.Interfaces.Shifts.IShiftManagementService.GetUrgentShiftsAsync:Humans.Domain.Entities.Shift
 Humans.Application.Interfaces.Shifts.IShiftSignupService.GetActiveSignupsForUserAsync:Humans.Domain.Entities.ShiftSignup
-Humans.Application.Interfaces.Shifts.IShiftSignupService.GetByShiftAsync:Humans.Domain.Entities.ShiftSignup
 Humans.Application.Interfaces.Shifts.IShiftSignupService.GetByUserAsync:Humans.Domain.Entities.ShiftSignup
 Humans.Application.Interfaces.Teams.ITeamService.GetAllTeamsAsync:Humans.Domain.Entities.Team
 Humans.Application.Interfaces.Teams.ITeamService.GetByIdsWithParentsAsync:Humans.Domain.Entities.Team
